requirements.txt
Flask==2.2.2
Flask-HTMLmin==2.2.1
gunicorn==20.0.4
PyYAML==6.0
requests==2.28.2
xmltodict==0.13.0
python-dateutil==2.8.2
EbookLib==0.18
tqdm==4.64.1
cssselect==1.2.0
ruamel.yaml==0.17.21
pymongo==3.11
dnspython==2.3
python-dotenv>=0.19.2
authlib>=1.0
beautifulsoup4==4.12.2
PyJWT==2.8.0
email-validator
auth0-python
